[Fix] admin / client / domain으로 패키지 분리
