+++ src/main/java/com/daall/howtoeat/domain/pt/Gym.java
@@ -1,4 +1,4 @@
-package com.daall.howtoeat.pt;
+package com.daall.howtoeat.domain.pt;
 
 import com.daall.howtoeat.common.Timestamped;
 import jakarta.persistence.*;
@@ -8,16 +8,12 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor
-@Table(name = "trainers")
-public class Trainer extends Timestamped {
+@Table(name = "gyms")
+public class Gym extends Timestamped {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "gym_id", nullable = false)
-    private Gym gym;
-
-    @Column( nullable = false)
+    @Column(unique = true, nullable = false)
     private String name;
 }
