+++ src/main/java/com/daall/howtoeat/domain/user/User.java
@@ -1,4 +1,4 @@
-package com.daall.howtoeat.user;
+package com.daall.howtoeat.domain.user;
 
 import com.daall.howtoeat.common.Timestamped;
 import com.daall.howtoeat.common.enums.Gender;
